[refactor] parser.js를 사이트별 모듈로 분리
src/lib/parser/
├── index.js            # 진입점: onBindParsingType, resolveSiteType, getExtractorBySiteType, getRandomUserAgent
├── http.js             # axios + userAgents + getHtml
├── utils.js            # getPriceWithoutString, emptyResult
└── extractors/
    ├── general.js
    ├── musinsa.js
    ├── wconcept.js
    ├── naver.js
    ├── cos.js
    ├── gmarket.js
    └── seoulstore.js

각 extractor 파일은 HTML 문자열을 받아 결과를 반환하는 순수 함수 1개만 export.
사이트 추가 시 extractors/<site>.js 추가 + index.js 라우팅 테이블에 항목 추가.

require 경로 호환: itemController.js, headlessParser.js의 require('../lib/parser')는
Node 디렉토리 require 규칙으로 자동으로 index.js를 찾으므로 변경 없음.

기능 변경 없음. Gmarket 가격 추출 버그는 별도 커밋으로 분리.
